Make sqrt call unambiguous

+++ stat.cpp
@@ -35,7 +35,7 @@ ci_t normal_sample::ci(double alpha) const {
     if (freedom() < 2) return std::make_pair(-std::numeric_limits<double>::infinity(), std::numeric_limits<double>::infinity());
     boost::math::students_t dist(freedom());
     double T = quantile(complement(dist, alpha / 2));
-    double w = T * stddev() / sqrt(n());
+    double w = T * stddev() / sqrt(static_cast<double>(n()));
     return std::make_pair(mean() - w, mean() + w);
 }
